Changed formColor to be slightly lighter. App no longer creates 'miscellaneous' category upon first startup; this was mostly done in the CustomContainer file. Added new view file, AddCategoryView which adds a button with an alert modifier for adding a new category. Used AddCategoryView in BrowseView. Added delete item button in EditFormView. Switched image file type to JPG from PNG for better file compression in FormView and EditItemView. Added delete banner notification to EditFormView deleteItem(). Switched list in ItemListView back to the default list to fix an animation bug that froze the screen for a few miliseconds after deleting an item. ItemListView appearance was also changed. ItemCardView now has the entire card background set to ultrathinmaterial instead of the vstack where the texts are housed.

+++ WheresMyStuff/View/BrowseView.swift
@@ -26,51 +26,54 @@ struct BrowseView: View {
                 .ignoresSafeArea(.all)
                 
                 VStack {
-//                    Text("Browse")
-//                        .foregroundStyle(Color.white)
-                    
                     //These show the recently viewed and added items views
                     RecentsCards()
                     
                     //lists the categories
                     Text("Categories")  //Consider making this a tab selection view to choose to browse between items and categories
-                        .padding(.top)
+                        .padding([.top, .bottom])
                         .foregroundStyle(Color.white)
                     
                     //MARK: category list
-                    List{
-//                        ForEach(categories[0].categoryList, id:\.self){ cat in
-                        ForEach(categories, id: \.self){ cat in
+                    if(!categories.isEmpty){
+                        List{
+                            ForEach(categories, id: \.self){ cat in
+                                
+                                NavigationLink (cat.name){
+                                    CategoryItemsListView(chosenCategory: cat.name)
+                                }
+                                .listRowSeparatorTint(Color.white)
+                                .listRowBackground(Color.clear)
+//                                .listRowBackground(border(Color.red))
+                            }
+                            .onDelete{ indexSet in
+                                for index in indexSet{
+                                    modelContext.delete(categories[index])
+                                }
+                            }
                             
-                            NavigationLink (cat.name){
-                                CategoryItemsListView(chosenCategory: cat.name)
-                            }
-                            .listRowSeparatorTint(Color.white)
-                            .listRowBackground(Color.clear)
-                            .listRowBackground(border(Color.white))
-                        }
-                        .onDelete{ indexSet in
-                            for index in indexSet{
-//                                categories[0].categoryList.remove(at: index)        //deletes the category
-                                modelContext.delete(categories[index])
-                            }
-                        }
-                        
-                    }// end list
-                    .listStyle(.plain)  //without this, the first row starts too low
-                    .scrollContentBackground(.hidden)
-                    .background(.ultraThinMaterial)
-                    .background(Color.clear)
-                    .cornerRadius(20)
-                    .padding([.leading, .trailing, .bottom], 38)
-                    
+                        }// end list
+                        .listStyle(.plain)  //without this, the first row starts too low
+//                        .scrollContentBackground(.hidden)
+                        .background(.ultraThinMaterial)
+                        .background(Color.clear)
+                        .cornerRadius(20)
+                        .padding([.leading, .trailing, .bottom], 38)
+                    } else {
+                        //MARK: Add category button
+                        AddCategoryView()
+                            .padding(100)
+                        .background(.ultraThinMaterial)
+                        .cornerRadius(20)
+                        Spacer()
+                    }
                 }//end vstack
                 .navigationTitle("Browse")
                 .navigationBarTitleDisplayMode(.inline)
             }//end zstack
         }//end navigation stack
     }//end body
-}//end body
+}
 
 #Preview {
     let container = try! ModelContainer(for: CategoryDataModel.self, ItemDataModel.self, configurations: ModelConfiguration(isStoredInMemoryOnly: true))
